Update jlcode to have the signature match jlsignature
This makes the return type of the method explicit, but more
importantly it changes the return type compared to previous behaviour.
Previously the method returned the first argument, now it returns the
return value from arb. This has the drawback of no longer returning
the modified value for in-place methods, which is different from what
Julia does in general. However it has the benefit of now actually
working for arb functions where the returned values is relevant, like
predicates.

The reason jlsignature is not used directly is that I could not figure
out how to get it to work correctly.