Update show methods to use arbcall methods when possible
Due to name clashes the standard print functions in Arb are not loaded
with arbcall and instead handled manually here.

The string return by Arb in dump_string is currently not freed
correctly. The commented out version doesn't work due to
LoadFlint.libflint not being a constant. This should be fixed.