Switch all occurrences of Clong/Culong to Int/UInt
The package have so far mainly been tested on Linux and Mac where
Clong corresponds to Int64, in Windows Clong corresponds to Int32
which didn't work with Arb.

Arb uses the type slong which is 32-bits on 32-bit systems and 64-bits
on 64-bit systems. This is the same guarantee as given by the Int type
in Julia, so using that should work in all cases.