Fix scalar addition and multiplication for polynomials
We recently changed the scalar addition and subtraction methods to use
inplace arithmetic to avoid allocations. This turned out to have
problem for the zero polynomial (the first coefficient is not
allocated) and it didn't properly normalise the polynomial afterwards.
This commit fixes both of these issues and adds tests for them.