Add specialised polynomial constructor for two-tuples
This is a common case, in particular when computing Taylor series
since you then often want to construct series on the form
ArbSeries((x, 1)). This specialised constructor removes the drawback
of having a tuple with elements of different types. In the case
ArbSeries((x::Arb, 1)) it reduces the number of allocations from 5 to
2 compared to the generic implementation.