Fix implementation of minimum and maximum
This addresses some of the issues discussed in
https://github.com/JuliaLang/julia/issues/45932.

For Julia 1.8.0 it completely fixes minimum and maximum, for earlier
versions of Julia it only partially fixes some of the issues.

The change completely removes the specialised implementation of
extrema. It also completely removes the specialised implementation of
minimum and maximum for Mag and Arf. The default implementations work
well, though are slower due to more allocations. In the future it
could maybe be an option to have a faster version using
MutableArithmetics.jl.

The Base implementation of minimum and maximum is fixed for Arb by
overloading Base._fast to make it correct for Arb. Before 1.8.0 this
is not enough to fix all problems and it thus keeps the specialised
implementation for Julia before 1.8.0-rc3 (latest release candidate as
of writing), for later versions the specialised implementation can be
removed.