Cleanup implementations of vectors and matrices, avoid use of XLike
Removes most occurrences of XLike types. It now only defines the most
basic methods directly on the structs.

Refactors the code for better reuse of code and clearer distinction of
struct types. Several style changes as well.

Removes the copyto! method for matrices. It should not be commonly
used for matrices since it just threats them as vectors anyway. Added
dimension checks for copy and copy! as well as tests.