Update constructors for vectors and matrices
This removes the default constructor for the types in favour for ones
which take the precision as a keyword argument.

It adds the shallow keyword for constructors from XLike types which
allows for shallow copies where they share the underlying Arb-struct.
Previously it always made a shallow copy with no option of making a
non-shallow copy. As a consequence the setprecision methods now copy
the input.

The tests have been updated to reflect that it defaults to a shallow
copy.