Improve conversion to Float16/32/64, BigFloat and Complex
It now correctly handles rounding in all cases and allows specifying
the precision for BigFloat.

Note that for BigFloat it uses different default values than the
constructors in Base. I always defaults to RoundNearest and uses the
precision given by the input argument. This means that it doesn't
depend on the global rounding mode and precision.

For Complex it now supports conversion from Acb to any Complex{T}, as
long as convert(T, ::ArbRef) is defined.