Update show implementation for Mag, Arf, Arb and Acb
It now defines `string` methods for the above types, instead of
directly defining `show`. The string methods take as keyword arguments
the possible flags to the printing functions in Flint.

Arf now uses the function `arf_get_string` instead of converting to
BigFloat. This allows handling numbers with exponents outside the
range for BigFloat.

The `string_nice` methods have been removed, in favour of the above
string methods. In fact the string_nice never worked for Mag and Arf.