ArbCall: Add special handling of "_series" functions
Flint has a number of functions for computing truncated series
expansions, these generally have a name ending in "_series". This
commit adds special handling of these functions.

In addition to the methods that were previously constructed for these
functions we now add another method. The new method removes the
"_series" suffix from the method name, restricts the input type to
only series types (and not polynomial types) and takes the default
length of the computed series from the first argument.

The main motivation for the new methods is to make it easier to write
generic code using mutability. For example now `Arblib.sin!` and be
applied to both `Arb` and `ArbSeries`, before one had to use
`Arblib.sin!` for `Arb` and `Arblib.sin_series!` for `ArbSeries`.